requirements.txt
openai
Pillow
requests
pyyaml
langchain
python-dotenv